Rename `css_plaquette.py` -> `css.py`
