+++ src/tqec/plaquette/library/css_test.py
